% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDrugSynonym.R
\name{checkDrugSynonym}
\alias{checkDrugSynonym}
\title{Check for drug synonyms}
\usage{
checkDrugSynonym(drugVector)
}
\arguments{
\item{drugVector}{A string vector of undefined length, with Drug names}
}
\value{
A dataframe containing: the input drug name, if it is approved or not, drughelper ID and a proper-more used synonym.
}
\description{
Given an input of drug synonyms, check if those drugs are approved and find a proper more used synonym
}
\examples{
\donttest{
checkDrugSynonym(c("Procaine", "Furazosin", "Embelin"))
}
}
