% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadAbsentFile.R
\name{downloadAbsentFile}
\alias{downloadAbsentFile}
\title{Download data from Chembl}
\usage{
downloadAbsentFile(dir = tempdir())
}
\arguments{
\item{dir}{Name of the directory where data is downloaded}
}
\value{
A dataset of drugs downloaded in current temp directory.
}
\description{
If it has not been downloaded yet, downloads data of drugs
}
