% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formattingDrugName.R
\name{formattingDrugName}
\alias{formattingDrugName}
\title{Normalization of drug names}
\usage{
formattingDrugName(DrugName)
}
\arguments{
\item{DrugName}{A string with the name of the drug to be normalized}
}
\value{
A string containing only alphanumeric content of the input drug and capitalized
}
\description{
This function corrects and capitalizes the names of a given vector of drugs.
}
\examples{
formattingDrugName("morphine")
}
