% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{InternalFunctions}
\alias{InternalFunctions}
\alias{tsv2rdata}
\title{Drughelper Internal Functions}
\usage{
tsv2rdata()
}
\value{
Internal outputs
}
\description{
Drughelper Internal Functions
}
\keyword{internal}
