% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drughelper-package.R
\docType{package}
\name{{drughelper}}
\alias{{drughelper}}
\alias{_PACKAGE}
\alias{{drughelper-package}}
\title{{drughelper}}
\description{
{
  To learn more about drughelper, start with the vignettes:
    \code{browseVignettes(package = "drughelper")}
}
}
\author{
{
  \strong{Maintainer}: Javier Garcia  \email{a905125@alumni.unav.es}

  Authors:
    \itemize{
      \item Javier Garcia
      \item Fernando Carazo
    }

  Other contributors:
    \itemize{
      \item Tecnun [copyright holder]
    }

}
}
\keyword{check}
\keyword{internal}
