% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{InternalFunctions}
\alias{InternalFunctions}
\alias{gittsv}
\alias{updateTable}
\title{Drughelper Internal Functions}
\usage{
gittsv()

updateTable(dhdrugs)
}
\value{
Internal outputs, called for side effects
}
\description{
Drughelper Internal Functions
}
\keyword{internal}
